ppt_generator/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
distro==1.9.0
Django==5.1.4
django-environ==0.11.2
django-ratelimit==4.1.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.3.0
openai==1.57.0
pillow==11.0.0
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic_core==2.27.1
PyJWT==2.10.1
python-dotenv==0.21.1
python-pptx==1.0.2
redis==5.2.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
sniffio==1.3.1
sqlparse==0.5.2
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
XlsxWriter==3.2.0
